// frontend/clients/generated/instructions/checkLiquidation.ts
export type CheckLiquidationInstructionAccounts = {
  liquidator: Signer;
  targetOwner: PublicKey | Pda;
  market: PublicKey | Pda;
  assetRegistry?: PublicKey | Pda;
  riskRegistry?: PublicKey | Pda;
  priceCache?: PublicKey | Pda;
  targetObligation?: PublicKey | Pda;
  systemProgram?: PublicKey | Pda;
};