// frontend/clients/generated/instructions/initPool.ts
export type InitPoolInstructionAccounts = {
  payer?: Signer;
  /** Admin = market.authority */
  authority?: Signer;
  market?: PublicKey | Pda;
  assetRegistry?: PublicKey | Pda;
  pool: PublicKey | Pda;
  mint: PublicKey | Pda;
  vaultAuth?: PublicKey | Pda;
  vault?: PublicKey | Pda;
  tokenProgram?: PublicKey | Pda;
  systemProgram?: PublicKey | Pda;
};

// Data.
export type InitPoolInstructionData = {
  discriminator: Uint8Array;
  rate: RateModel;
  mint: PublicKey;
};

export type InitPoolInstructionDataArgs = {
  rate: RateModelArgs;
  mint: PublicKey;
};
// ...
export type InitPoolInstructionArgs = InitPoolInstructionDataArgs;