// frontend/clients/generated/instructions/leverageExistingDeposit.ts
export type LeverageExistingDepositInstructionAccounts = {
  owner: Signer;
  market: PublicKey | Pda;
  authority?: PublicKey | Pda;
  assetRegistry?: PublicKey | Pda;
  riskRegistry?: PublicKey | Pda;
  priceCache?: PublicKey | Pda;
  borrowPool: PublicKey | Pda;
  depositPool: PublicKey | Pda;
  obligation?: PublicKey | Pda;
  systemProgram?: PublicKey | Pda;
};

// Data.
export type LeverageExistingDepositInstructionData = {
  discriminator: Uint8Array;
  borrowAmount: bigint;
};

export type LeverageExistingDepositInstructionDataArgs = {
  borrowAmount: number | bigint;
};
// ...
export type LeverageExistingDepositInstructionArgs =
  LeverageExistingDepositInstructionDataArgs;