// frontend/clients/generated/instructions/registerAsset.ts
export type RegisterAssetInstructionAccounts = {
  authority?: Signer;
  market?: PublicKey | Pda;
  assetRegistry?: PublicKey | Pda;
  riskRegistry?: PublicKey | Pda;
  systemProgram?: PublicKey | Pda;
};

// Data.
export type RegisterAssetInstructionData = {
  discriminator: Uint8Array;
  mint: PublicKey;
  decimals: number;
  pythPrice: PublicKey;
  pythFeedId: Option<Uint8Array>;
  enabledAsCollateral: boolean;
};

export type RegisterAssetInstructionDataArgs = {
  mint: PublicKey;
  decimals: number;
  pythPrice: PublicKey;
  pythFeedId: OptionOrNullable<Uint8Array>;
  enabledAsCollateral: boolean;
};
// ...
export type RegisterAssetInstructionArgs = RegisterAssetInstructionDataArgs;