// frontend/clients/generated/instructions/repay.ts
export type RepayInstructionAccounts = {
  owner: Signer;
  market?: PublicKey | Pda;
  authority?: PublicKey | Pda;
  assetRegistry?: PublicKey | Pda;
  pool: PublicKey | Pda;
  ownerTokenAta: PublicKey | Pda;
  poolVault: PublicKey | Pda;
  vaultAuth?: PublicKey | Pda;
  obligation?: PublicKey | Pda;
  tokenProgram?: PublicKey | Pda;
  systemProgram?: PublicKey | Pda;
};

// Data.
export type RepayInstructionData = {
  discriminator: Uint8Array;
  amount: bigint;
};

export type RepayInstructionDataArgs = { amount: number | bigint };
// ...
export type RepayInstructionArgs = RepayInstructionDataArgs;