// frontend/clients/generated/instructions/setRiskPair.ts
export type SetRiskPairInstructionAccounts = {
  authority?: Signer;
  market?: PublicKey | Pda;
  assetRegistry?: PublicKey | Pda;
  riskRegistry?: PublicKey | Pda;
};

// Data.
export type SetRiskPairInstructionData = {
  discriminator: Uint8Array;
  aMint: PublicKey;
  bMint: PublicKey;
  ltvBps: number;
  liqThresholdBps: number;
  liqBonusBps: number;
};

export type SetRiskPairInstructionDataArgs = {
  aMint: PublicKey;
  bMint: PublicKey;
  ltvBps: number;
  liqThresholdBps: number;
  liqBonusBps: number;
};
// ...
export type SetRiskPairInstructionArgs = SetRiskPairInstructionDataArgs;