// frontend/clients/generated/instructions/updatePrices.ts
export type UpdatePricesInstructionAccounts = {
  authority?: Signer;
  market?: PublicKey | Pda;
  assetRegistry?: PublicKey | Pda;
  priceCache?: PublicKey | Pda;
};

// Data.
export type UpdatePricesInstructionData = {
  discriminator: Uint8Array;
  updates: Array<PriceUpdate>;
};

export type UpdatePricesInstructionDataArgs = {
  updates: Array<PriceUpdateArgs>;
};
// ...
export type UpdatePricesInstructionArgs = UpdatePricesInstructionDataArgs;