// frontend/clients/generated/instructions/withdraw.ts
export type WithdrawInstructionAccounts = {
  owner: Signer;
  market?: PublicKey | Pda;
  authority?: PublicKey | Pda;
  assetRegistry?: PublicKey | Pda;
  riskRegistry?: PublicKey | Pda;
  priceCache?: PublicKey | Pda;
  pool: PublicKey | Pda;
  poolVault: PublicKey | Pda;
  vaultAuth?: PublicKey | Pda;
  ownerTokenAta: PublicKey | Pda;
  obligation?: PublicKey | Pda;
  tokenProgram?: PublicKey | Pda;
  systemProgram?: PublicKey | Pda;
};

// Data.
export type WithdrawInstructionData = {
  discriminator: Uint8Array;
  amount: bigint;
};

export type WithdrawInstructionDataArgs = { amount: number | bigint };
// ...
export type WithdrawInstructionArgs = WithdrawInstructionDataArgs;